redis初始化（区分cache和biz） cacheRedis编写接口，只适用获取普通缓存信息

+++ pkg/types/types.go
@@ -25,10 +25,10 @@ type SqlSourceConf struct {
 type BizRedisConf struct {
 	Addr     string `json:",omitempty"`
 	Password string `json:",omitempty"`
-	DB       int8   `json:",omitempty"`
+	DB       int    `json:",omitempty"`
 }
 type CacheRedisConf struct {
 	Addr     string `json:",omitempty"`
 	Password string `json:",omitempty"`
-	DB       int8   `json:",omitempty"`
+	DB       int    `json:",omitempty"`
 }
